feat(reestructure): make changes in all project
