+++ index.html
@@ -14,7 +14,7 @@
 </head>
 <body>
     <h1>RxJS Demo</h1>
-    Check the console.logs renders above;
+    <p>Check the console.logs renders above</p>
     <div>
       <ul id="list"></ul>
     </div>
